feat(installer): opt-in to font installation

+++ install.sh
@@ -1,24 +1,26 @@
 #!/bin/bash -e
 
 USER_RC_PATH=${1:-"${HOME}"}
+SKIP_FONTS=${2:-"true"}
 
 if [ -x "$(command -v git)" ] && [ ! -d "${USER_RC_PATH}/.tmux/plugins/tpm" ]; then
 	git clone https://github.com/tmux-plugins/tpm "${USER_RC_PATH}/.tmux/plugins/tpm"
 fi
 
-font_dir="${USER_RC_PATH}/.local/share/fonts/"
-if [ ! -f "${font_dir}Hack Regular Nerd Font Complete.ttf" ]; then
-	mkdir -p $font_dir
-	curl -fLo ${USER_RC_PATH}/Hack.zip https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/Hack.zip
-	unzip -d ${font_dir} ${USER_RC_PATH}/Hack.zip
-	rm  ${USER_RC_PATH}/Hack.zip
+if [ "${USER_RC_PATH}" != "true" ]; then
+	font_dir="${USER_RC_PATH}/.local/share/fonts/"
+	if [ ! -f "${font_dir}Hack Regular Nerd Font Complete.ttf" ]; then
+		mkdir -p $font_dir
+		curl -fLo ${USER_RC_PATH}/Hack.zip https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/Hack.zip
+		unzip -d ${font_dir} ${USER_RC_PATH}/Hack.zip
+		rm  ${USER_RC_PATH}/Hack.zip
 
-	# Reset font cache on Linux
-	if [[ -n $(command -v fc-cache) ]]; then
-	  fc-cache -f "$font_dir"
+		# Reset font cache on Linux
+		if [[ -n $(command -v fc-cache) ]]; then
+		  fc-cache -f "$font_dir"
+		fi
 	fi
 fi
-
 
 
 # list of files/folders to symlink in homedir
